e2e: move validation to validate package and refactor calls
With this change util/validations.go is moved separate
package under validate/validation.go, which will be also
used in ramenctl.
